requirements.txt
google-adk
google-cloud-bigquery
google-cloud-aiplatform
google-generativeai
fastapi
uvicorn
pydantic
python-dotenv
pandas
pyyaml
python-multipart
requests
